@@ -1,22 +1,36 @@
 using System;
+using System.Net;
 using System.Net.Http;
+using System.Reflection.Metadata;
 using System.Text;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Options;
 
 namespace Celigo.ServiceManager.NetSuite.REST
 {
+    public struct Passport
+    {
+        public string Account { get; set; }    
+        public string Email { get; set; }
+        public string Password { get; set; }
+    }
+    
     public interface IRestletClient
     {
-        Task<HttpResponseMessage> Get(string account, string token, string tokenSecret, params (string key, string value)[] queryParams);
-        Task<HttpResponseMessage> Post<T>(string account, string token, string tokenSecret, T message, params (string key, string value)[] queryParams);
+        Task<HttpResponseMessage> Get(in string account, in string token, in string tokenSecret, params (string key, string value)[] queryParams);
+        Task<HttpResponseMessage> Post<T>(in string account, in string token, in string tokenSecret, in T message, params (string key, string value)[] queryParams);
+        
+        Task<HttpResponseMessage> Get(in Passport passport, params (string key, string value)[] queryParams);
+        Task<HttpResponseMessage> Post<T>(in Passport passport, in T message, params (string key, string value)[] queryParams);
     }
 
     public class RestletClient: RestClient, IRestletClient
     {
+        private static readonly HttpContent _emptyContent = new StringContent("", Encoding.UTF8, "application/json");
+
         private readonly RestletConfig _restlet;
         
-        private const string _restletRelativePath = ".restlets.api.netsuite.com/app/site/hosting/restlet.nl?";
+        private const string _restletRelativePath = ".restlets.api.netsuite.com/app/site/hosting/restlet.nl";
         private const string _scriptParamName = "script";
         private const string _deployParamName = "deploy";
         
@@ -34,18 +48,49 @@ namespace Celigo.ServiceManager.NetSuite.REST
         public RestletClient(HttpClient httpClient, IOptions<RestClientOptions> options, IOptions<RestletConfig> restlet)
             : this(httpClient, options, restlet.Value) { }
 
-        public Task<HttpResponseMessage> Get(string account, string token, string tokenSecret, params (string key, string value)[] queryParams)
+        public Task<HttpResponseMessage> Get(in string account, 
+                                             in string token, 
+                                             in string tokenSecret, 
+                                             params (string key, string value)[] queryParams)
         {
-            var (requestUrl, authHeader) = GenerateHeaderAndUrl(account, token, tokenSecret, queryParams);
-            return this.SendRequest(HttpMethod.Get, requestUrl, authHeader);
+            var (requestUrl, authHeader) = GenerateHeaderAndUrl(account, token, tokenSecret, isBasicAuth: false, queryParams);
+            return this.SendRequest(HttpMethod.Get, requestUrl, authHeader, _emptyContent);
         }
 
-        public Task<HttpResponseMessage> Post<T>(string account, string token, string tokenSecret, T message, params (string key, string value)[] queryParams)
+        public Task<HttpResponseMessage> Post<T>(in string account, 
+                                                 in string token, 
+                                                 in string tokenSecret, 
+                                                 in T message, 
+                                                 params (string key, string value)[] queryParams)
         {
-            var (requestUrl, authHeader) = GenerateHeaderAndUrl(account, token, tokenSecret, queryParams);
+            var (requestUrl, authHeader) = GenerateHeaderAndUrl(account, token, tokenSecret, isBasicAuth: false, queryParams);
             return this.SendRequest(HttpMethod.Post, requestUrl, authHeader, CreateJsonMessageContent(message));
         }
-        
+
+        public Task<HttpResponseMessage> Get(in Passport passport, 
+                                             params (string key, string value)[] queryParams)
+        {
+            ValidateBasicCreds(passport);
+            
+            var (requestUrl, authHeader) = GenerateHeaderAndUrl(passport.Account, 
+                                                                passport.Email, 
+                                                                passport.Password, 
+                                                                isBasicAuth: true, 
+                                                                queryParams);
+            
+            return this.SendRequest(HttpMethod.Get, requestUrl, authHeader, _emptyContent);
+        }
+
+        public Task<HttpResponseMessage> Post<T>(in Passport passport,
+                                                 in T message, 
+                                                 params (string key, string value)[] queryParams)
+        {
+            ValidateBasicCreds(passport);
+            
+            var (requestUrl, authHeader) = GenerateHeaderAndUrl(passport.Account, passport.Email, passport.Password, isBasicAuth: true, queryParams);
+            return this.SendRequest(HttpMethod.Post, requestUrl, authHeader, CreateJsonMessageContent(message));
+        }
+
         private StringBuilder CreateUrlBuilder(string account) => new StringBuilder("https://")
                                                                 .Append(account.ToLowerInvariant().Replace('_', '-'))
                                                                 .Append(_restletRelativePath);
@@ -53,7 +98,7 @@ namespace Celigo.ServiceManager.NetSuite.REST
         private Uri CreateRequestUrl(StringBuilder urlBuilder, (string key, string value)[] queryParams)
         {
             urlBuilder
-                .Append("script=")
+                .Append("?script=")
                 .Append(_restlet.Script)
                 .Append("&deploy=")
                 .Append(_restlet.Deploy);
@@ -69,10 +114,15 @@ namespace Celigo.ServiceManager.NetSuite.REST
             return new Uri(urlBuilder.ToString());
         }
 
-        private (Uri requestUrl, string authHeader) GenerateHeaderAndUrl(string account, string token, string tokenSecret, (string key, string value)[] queryParams)
-        {
-            var urlBuilder = this.CreateUrlBuilder(account);
+        private string GetBasicAuthHeaderValue(string username, string password) =>
+            $"NLAuth nlauth_email={WebUtility.UrlEncode(username)}, nlauth_signature={Uri.EscapeDataString(password)}";
 
+        private (Uri requestUrl, string authHeader) GenerateHeaderAndUrl(string account, 
+                                                                         string key, 
+                                                                         string secret, 
+                                                                         bool isBasicAuth,
+                                                                         (string key, string value)[] queryParams)
+        {
             var oauthParams = this.GetCommonOAuthParameters();
             oauthParams.Add(_scriptParamName, _restlet.Script);
             oauthParams.Add(_deployParamName, _restlet.Deploy);
@@ -81,9 +131,21 @@ namespace Celigo.ServiceManager.NetSuite.REST
                 oauthParams.Add(queryParams[i].key, queryParams[i].value);
             }
 
-            string authHeader = this.GetAuthorizationHeaderValue(account, urlBuilder.ToString(), oauthParams, token, tokenSecret, "GET");
-            Uri requestUrl = this.CreateRequestUrl(urlBuilder, queryParams);
+            var urlBuilder = this.CreateUrlBuilder(account);
+            
+            string authHeader = isBasicAuth
+                ? this.GetBasicAuthHeaderValue(key, secret)
+                : this.GetAuthorizationHeaderValue(account, urlBuilder.ToString(), oauthParams, key, secret, "GET");
+
+            var requestUrl = this.CreateRequestUrl(urlBuilder, queryParams);
             return (requestUrl, authHeader);
+        }
+
+        private void ValidateBasicCreds(in Passport passport)
+        {
+            _ = passport.Account ?? throw new ArgumentNullException($"{nameof(Passport)}.{nameof(Passport.Account)}");
+            _ = passport.Email ?? throw new ArgumentNullException($"{nameof(Passport)}.{nameof(Passport.Email)}");
+            _ = passport.Password ?? throw new ArgumentNullException($"{nameof(Passport)}.{nameof(Passport.Password)}");
         }
     }
 }
